+++ Assets/Sounds/audioManager.cs
@@ -108,7 +108,7 @@ public class audioManager : MonoBehaviour
     // FindObjectOfType<audioManager>().play("dashSound"); example of calling in other scripts
 
     // change menu music volume on the fly
-    public void musicVolumeChange(int volume)
+    public void musicVolumeChange(float volume)
     {
         musicVolume = volume;
         sound musicClip = Array.Find(sounds, sound => sound.name == "menuMusic");
